Split layouts into includes
Also standardise on date for published at. Ruhoh (if I recall) borked on
date parsing, so I needed to include redundant data in each post's meta
data to hold the month name and all that crap. That's gone in this
change.

The url format has changed to /blog/:title. I might reintroduce the date
there, even though these urls are supposed to be perma. Oh well.